Added headers files and shared library option [make lib]
