+++ .gitignore
@@ -1,2 +1,6 @@
 scanner
+libscanner.so
+.*o
+.*d
+*.out
 .gitignore
