+++ external/inc/crc32c.h
@@ -1,0 +1,6 @@
+#ifndef __CRC32C_H
+    #define __CRC32C_H
+
+uint32_t calc_crc32c (char *data, ssize_t len);
+
+#endif
