+++ external/inc/winnowing.h
@@ -1,0 +1,9 @@
+#ifndef __WINNOWING_H
+    #define __WINNOWING_H
+
+#include <stdint.h>    
+
+uint32_t winnowing (char *src, uint32_t *hashes, uint32_t *lines, uint32_t limit);
+
+
+#endif
