+++ external/src/crc32c.c
@@ -223,7 +223,6 @@ static void crc32c_init_hw(void)
 }
 
 /* Compute CRC-32C using the Intel hardware instruction. */
-#ifndef CRC32_SOFTWARE_MODE
 static uint32_t crc32c_hw(uint32_t crc, const void *buf, size_t len)
 {
     const unsigned char *next = buf;
@@ -337,13 +336,7 @@ uint32_t crc32c(uint32_t crc, const void *buf, size_t len)
     SSE42(sse42);
     return sse42 ? crc32c_hw(crc, buf, len) : crc32c_sw(crc, buf, len);
 }
-#else
-uint32_t crc32c(uint32_t crc, const void *buf, size_t len)
-{
-      return crc32c_sw(crc, buf, len);
-}
 
-#endif //CRC32_SOFTWARE_MODE
 
 #define SIZE (262144*3)
 #define CHUNK SIZE
