+++ external/src/winnowing.c
@@ -4,7 +4,7 @@
  *
  * Winnowing algorithm implementation
  *
- * Copyright (C) 2018-2020 SCANOSS LTD
+ * Copyright (C) 2018-2020 SCANOSS.COM
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -22,12 +22,14 @@
 
 
 #include <stdbool.h>
-#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
 #include <openssl/md5.h>
-#include "crc32c.c"
+//#include "external/crc32c/crc32c.c"
+#include "crc32c.h"
+#include "winnowing.h"
+
 
 uint8_t GRAM  = 30;   // Winnowing gram size in bytes
 uint8_t WINDOW = 64;  // Winnowing window size in bytes
@@ -35,7 +37,7 @@ uint32_t MAX_UINT32 = 4294967295;
 
 /* Convert case to lowercase, and return zero if it isn't a letter or number
    Do it fast and independent from the locale configuration (avoid string.h) */
-uint8_t normalize (uint8_t byte)
+static uint8_t normalize (uint8_t byte)
 {
 	if (byte < '0')  return 0;
 	if (byte > 'z')  return 0;
@@ -46,7 +48,7 @@ uint8_t normalize (uint8_t byte)
 }
 
 /* Left shift one window */
-void shift_window(uint32_t *window)
+static void shift_window(uint32_t *window)
 {
 	for (uint32_t i = 0; i < (WINDOW - 1); i++)
 	{
@@ -56,7 +58,7 @@ void shift_window(uint32_t *window)
 }
 
 /* Left shift one gram */
-void shift_gram(uint8_t *gram)
+static void shift_gram(uint8_t *gram)
 {
 	for (uint32_t i = 0; i < (GRAM - 1); i++)
 	{
@@ -66,7 +68,7 @@ void shift_gram(uint8_t *gram)
 }
 
 /* Select smaller hash for the given window */
-uint32_t smaller_hash(uint32_t *window)
+static uint32_t smaller_hash(uint32_t *window)
 {
 	uint32_t hash = MAX_UINT32;
 	for (uint32_t h = 0; h < WINDOW; h++)
@@ -78,7 +80,7 @@ uint32_t smaller_hash(uint32_t *window)
 
 /* Add the given "hash" to the "hashes" array and the corresponding "line" to the "lines" array
    updating the hash counter and returning the last added hash */
-uint32_t add_hash(uint32_t hash, uint32_t line, uint32_t *hashes, uint32_t *lines, uint32_t last, uint32_t *counter)
+static uint32_t add_hash(uint32_t hash, uint32_t line, uint32_t *hashes, uint32_t *lines, uint32_t last, uint32_t *counter)
 {
 
 	/* Consecutive repeating hashes are ignored */
