+++ src/scanner.c
@@ -26,24 +26,16 @@
 #include <openssl/md5.h>
 #include <openssl/ssl.h>
 #include <dirent.h> 
-#include <stdbool.h>
 #include <stdint.h>
 #include <stdio.h>
 #include <string.h>
 #include <sys/stat.h>
 
-#include "blacklist_ext.h"
-#include "external/winnowing.c"
 
-#define VERSION "1.1.0"
-#define API_HOST "osskb.org"
-#define API_PORT "443"
-#define MAX_HEADER_LEN 1048576
-#define BUFFER_SIZE_MAX 524288
-#define BUFFER_SIZE_MIN 256
-#define BUFFER_SIZE_DEFAULT 524288
-#define MAX_FILE_SIZE (1024 * 1024 * 4)
-#define MIN_FILE_SIZE 128
+#include "scanner.h"
+#include "blacklist_ext.h"
+#include "winnowing.h"
+
 
 //#define DEBUG /* uncomment this define to enable some debug outputs */
 
@@ -310,72 +302,4 @@ bool dir_proc(char * path)
     
     closedir(d); 
     return true;
-}
-
-
-int main(int argc, char *argv[])
-{
-    int param = 0;
-    
-    /* Command parser */
-    while ((param = getopt (argc, argv, "f:b:hv")) != -1)
-        switch (param)
-        {
-        case 'b': //set http response buffe size
-            buffer_size = atoi(optarg);
-            if (buffer_size > BUFFER_SIZE_MAX || buffer_size < BUFFER_SIZE_MIN )
-            {
-                buffer_size = BUFFER_SIZE_DEFAULT;
-                fprintf(stderr,"Wrong buffer size, using default: %d",buffer_size);
-            }
-            
-            break;
-        case 'v':
-            verbose = true;
-             fprintf(stderr, "verbose mode\n");
-            break;
-        case 'f':
-            if (strstr(optarg,"plain") || strstr(optarg,"spdx") || strstr(optarg,"cyclonedx"))
-                strncpy(format,optarg,sizeof(format));
-            else
-                fprintf(stderr, "%s is not a valid output format, using plain\n",optarg);
-                
-            if (verbose) fprintf(stderr, "Selected format: %s\n",format);
-            break;
-        case 'h':
-        default:
-            fprintf(stderr, "SCANOSS scanner-%s\n", VERSION);
-            fprintf(stderr, "Usage: scanner FILE or scanner DIR\n");
-            fprintf(stderr, "Option\t\t Meaning\n");
-            fprintf(stderr, "-h\t\t Show this help\n");
-            fprintf(stderr, "-f<format>\t Output format, could be: plain (default), spdx, spdx_xml or cyclonedx.\n");
-            fprintf(stderr, "-b<bytes>\t HTTP response buffer size, default: %d bytes\n",BUFFER_SIZE_DEFAULT);
-            fprintf(stderr, "-v\t\t Enable verbosity (via STDERR)\n");
-            fprintf(stderr, "\nFor more information, please visit https://scanoss.com\n");
-            exit(EXIT_FAILURE);
-           break;
-        }
-    
-       
-    char *path = argv[optind];
-    
-    if (is_file(path))
-    {
-        file_proc(path);
-    }
-    else if (is_dir(path)) 
-    {
-        int path_len = strlen(path);
-        if (path_len > 1 && path[path_len-1] == '/') //remove extra /
-            path[path_len-1] = '\0';
-        
-        dir_proc(path);
-    }
-    else
-    {
-        fprintf(stderr,"%s is not a file\n", path);
-        return EXIT_FAILURE;
-    }
-	
-	return EXIT_SUCCESS;
 }
